refactor more to per-bit files
